+++ lua/plugins/config_voterandom.lua
@@ -6,6 +6,7 @@ local function SetupDefaultConfig()
 	kDAKConfig.VoteRandom = { }
 	kDAKConfig.VoteRandom.kVoteRandomInstantly = false
 	kDAKConfig.VoteRandom.kVoteRandomAlwaysEnabled = false
+	kDAKConfig.VoteRandom.kVoteRandomOnGameStart = false
 	kDAKConfig.VoteRandom.kVoteRandomDuration = 30
 	kDAKConfig.VoteRandom.kVoteRandomMinimumPercentage = 60
 	kDAKConfig.VoteRandom.kVoteRandomChatCommands = { "voterandom", "random" }